pyproject.toml
[project]
name = "ai-tools"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "absl-py==2.3.1",
    "astunparse==1.6.3",
    "beautifulsoup4==4.13.4",
    "blinker==1.9.0",
    "cachetools==5.5.2",
    "certifi==2025.6.15",
    "charset-normalizer==3.4.2",
    "click==8.2.1",
    "colorama==0.4.6",
    "deepface==0.0.93",
    "fastapi[standard]>=0.115.14",
    "filelock==3.18.0",
    "fire==0.7.0",
    "flask==3.1.1",
    "flask-cors==6.0.1",
    "flatbuffers==25.2.10",
    "gast==0.6.0",
    "gdown==5.2.0",
    "google-auth==2.40.3",
    "google-auth-oauthlib==1.0.0",
    "google-pasta==0.2.0",
    "grpcio==1.73.1",
    "gunicorn==23.0.0",
    "h5py==3.14.0",
    "idna==3.10",
    "itsdangerous==2.2.0",
    "jinja2==3.1.6",
    "joblib==1.5.1",
    "keras==2.14.0",
    "libclang==18.1.1",
    "lz4==4.4.4",
    "markdown==3.8.2",
    "markupsafe==3.0.2",
    "ml-dtypes==0.2.0",
    "mtcnn==1.0.0",
    "numpy==1.26.4",
    "oauthlib==3.3.1",
    "opencv-python==4.11.0.86",
    "opt-einsum==3.4.0",
    "packaging==25.0",
    "pandas==2.3.0",
    "pillow==11.3.0",
    "protobuf==4.25.8",
    "pyasn1==0.6.1",
    "pyasn1-modules==0.4.2",
    "pysocks==1.7.1",
    "python-dateutil==2.9.0.post0",
    "python-multipart>=0.0.20",
    "pytz==2025.2",
    "requests==2.32.4",
    "requests-oauthlib==2.0.0",
    "retina-face==0.0.17",
    "rsa==4.9.1",
    "setuptools==80.9.0",
    "six==1.17.0",
    "soupsieve==2.7",
    "tensorboard==2.14.1",
    "tensorboard-data-server==0.7.2",
    "tensorflow==2.14.0",
    "tensorflow-estimator==2.14.0",
    "tensorflow-intel==2.14.0",
    "tensorflow-io-gcs-filesystem==0.31.0",
    "termcolor==3.1.0",
    "torch>=2.7.1",
    "tqdm==4.67.1",
    "transformers>=4.53.1",
    "typing-extensions==4.14.1",
    "tzdata==2025.2",
    "urllib3==2.5.0",
    "uvicorn[standard]==0.19.0",
    "werkzeug==3.1.3",
    "wheel==0.45.1",
    "wrapt==1.14.1",
]
